+++ jupiterContext/Cart.cs
@@ -30,7 +30,7 @@ namespace jupiterCore.jupiterContext
         public byte? IsEmailSend { get; set; }
         public byte? IsPickup { get; set; }
         public string Region { get; set; }
-        public byte? IsPay { get; set; }
+        public int? IsPay { get; set; }
         public byte? IsExpired { get; set; }
         public int? CartStatusId { get; set; }
         public decimal? RentalPaidFee { get; set; }
